requirements.txt
altgraph==0.17
future==0.18.2
lxml==4.5.2
numpy==1.19.1
pandas==1.0.5
pefile==2019.4.18
pendulum==2.1.1
PyInstaller==3.6
pyodbc==4.0.30
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
pytzdata==2020.1
pywin32-ctypes==0.2.0
six==1.15.0
SQLAlchemy==1.3.18
xlrd==1.2.0
